Add support to the "map" validator for constraining the keys (#95)
* Add support for a "key" constraint to the "map" validator

This constraint allows a validator to be applied to the keys of the map, in
addition to the usual validator applied to the values

* Add some tests of the new "key" constraint on "map"

These tests cover the case of a map at the base level, as well as a map in a
nested list. There are much more complicated cases that could, and perhaps
ought, to be tested though...

* Minor cosmetic fix

* Update documentation of "Map" to describe the "key" constraint

* Add keywords to the documented "List" prototype

* Remove unusual extra line breaks

Utterly minor thing, but the document is otherwise bereft of double blank lines
so if there is a consistent "style" then that style is to only use one blank
line to separate sections.