Prevent int and num validators from accepting bool values. (#109)
Adds a change to the Integer and Numeric validators such that they
won't interpret boolean values (False, True) as valid, as well as tests
outlining the expected behavior (and that 0 and 1 are still numeric).

This is an issue caused by the origin of Bool in python, as outlined in
PEP 285 -- Adding a bool type (www.python.org/dev/peps/pep-0285/):
> The bool type would be a straightforward subtype (in C) of the int
> type, and the values False and True would behave like 0 and 1 in most
> respects

For historical reasons, bool is a subtype of int in Python, but that is
not the case in YAML (see e.g. https://yaml.org/type/bool.html). This
can cause issues where Yamale validates a YAML document with a bool
in an int-only field, and a downstream consumer (whether a non-python
program or a python program using .str() like e.g. jinja2) ends in an
invalid state.

The behavior of the current int and num validators can be emulated by
changing instances of 'int()' and 'num()' to 'any(int(), bool())' and
'any(num(), bool())', so there should be no loss of functionality for
users that rely on this type coercion.